# shared/contracts/openapi.nexus-core.snapshot.yaml
openapi: 3.0.3
info:
  title: Nexus Core API
  version: "1.1.0"
  description: |
    Multi-tenant Agent Tool Gateway for controlled tool execution.

    Authentication supports API key (`X-NEXUS-GATEWAY-KEY`) and optional JWT bearer mode (`Authorization: Bearer`) when enabled by feature flags.

    Idempotency semantics for `POST /v1/run` on WRITE tools:
    - Same `Idempotency-Key` + same request fingerprint + `COMPLETED` => replay cached success.
    - Same `Idempotency-Key` + same fingerprint + `IN_PROGRESS` => `IDEMPOTENCY_IN_PROGRESS`.
    - Same `Idempotency-Key` + different fingerprint => `IDEMPOTENCY_KEY_CONFLICT`.
    - Same `Idempotency-Key` + same fingerprint + `FAILED` => terminal replay of cached error.
    - To retry after `FAILED`, clients MUST use a new `Idempotency-Key`.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/tools/{name}/policies:
    get:
      tags: [Policies]
      operationId: listPoliciesByTool
      summary: List policies for a tool
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      responses:
        "200":
          description: Policies list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
    post:
      tags: [Policies]
      operationId: createPolicyByTool
      summary: Create policy for a tool
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCreateRequest"
      responses:
        "201":
          description: Policy created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/policies/{id}:
    put:
      tags: [Policies]
      operationId: updatePolicyById
      summary: Partial update policy by ID
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyUpdateRequest"
      responses:
        "200":
          description: Policy updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/tools/{name}/secrets:
    get:
      tags: [Secrets]
      operationId: listToolSecrets
      summary: List secret metadata (never returns plaintext values)
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      responses:
        "200":
          description: Secret metadata list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SecretMetadataResponse"
                required: [items]
        "403":
          description: Requires secops/admin role or admin:secrets scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Secrets]
      operationId: upsertToolSecret
      summary: Upsert encrypted secret
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretUpsertRequest"
      responses:
        "200":
          description: Secret metadata saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretMetadataResponse"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Secrets]
      operationId: deleteToolSecret
      summary: Delete secret by key_name
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
        - name: key_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted.
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/policy-proposals:
    get:
      tags: [Policies]
      operationId: listPolicyProposals
      summary: List policy proposals
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, approved, rejected, shadow]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Proposal list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyProposalListResponse"
    post:
      tags: [Policies]
      operationId: createPolicyProposal
      summary: Create policy proposal
      security:
        - NexusApiKey: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyProposalCreateRequest"
      responses:
        "201":
          description: Proposal created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyProposalItem"
  /v1/policy-proposals/{id}/approve:
    post:
      tags: [Policies]
      operationId: approvePolicyProposal
      summary: Approve policy proposal
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Proposal updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyProposalItem"
  /v1/policy-proposals/{id}/reject:
    post:
      tags: [Policies]
      operationId: rejectPolicyProposal
      summary: Reject policy proposal
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Proposal updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyProposalItem"
  /v1/policy-proposals/{id}/shadow:
    post:
      tags: [Policies]
      operationId: shadowPolicyProposal
      summary: Start proposal in shadow mode
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Proposal updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyProposalItem"
...
components:
  securitySchemes:
    NexusApiKey:
      type: apiKey
      in: header
      name: X-NEXUS-GATEWAY-KEY
      description: Tenant API key (stored hashed in DB).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional JWT+JWKS auth mode when enabled by feature flags.
  parameters:
    ToolNamePath:
      name: name
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorObject:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        error:
          $ref: "#/components/schemas/ErrorObject"
      required: [request_id, error]
...
    Policy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tool_id:
          type: string
          format: uuid
        effect:
          type: string
          enum: [allow, deny]
        priority:
          type: integer
        conditions:
          type: object
          additionalProperties: true
        limits:
          type: object
          additionalProperties: true
        reason_template:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, tool_id, effect, priority, conditions, limits, reason_template, enabled, created_at, updated_at]
    PolicyCreateRequest:
      type: object
      properties:
        effect:
          type: string
          enum: [allow, deny]
        priority:
          type: integer
        conditions:
          type: object
          additionalProperties: true
        limits:
          type: object
          additionalProperties: true
        reason_template:
          type: string
        enabled:
          type: boolean
      required: [effect, priority, conditions, limits, reason_template, enabled]
    PolicyUpdateRequest:
      type: object
      properties:
        effect:
          type: string
          enum: [allow, deny]
        priority:
          type: integer
        conditions:
          type: object
          additionalProperties: true
        limits:
          type: object
          additionalProperties: true
        reason_template:
          type: string
        enabled:
          type: boolean
    PolicyListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Policy"
      required: [items]
...
    SecretUpsertRequest:
      type: object
      properties:
        secret_type:
          type: string
          enum: [header, bearer]
        key_name:
          type: string
        value:
          type: string
          description: Plaintext provided by caller; encrypted at rest server-side.
        enabled:
          type: boolean
      required: [secret_type, value]
    SecretMetadataResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        secret_type:
          type: string
          enum: [header, bearer]
        key_name:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, secret_type, key_name, enabled, created_at, updated_at]
...
    PolicyProposalCreateRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, pending]
        diff:
          type: object
          additionalProperties: true
        rationale:
          type: string
        tests_suggested:
          type: array
          items:
            type: string
        rollback_plan:
          type: string
      required: [diff, rationale]
    PolicyProposalItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, pending, approved, rejected, shadow]
        diff:
          type: object
          additionalProperties: true
        rationale:
          type: string
        tests_suggested:
          type: array
          items:
            type: string
        rollback_plan:
          type: string
        created_by:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        decided_by:
          type: string
          nullable: true
        decided_at:
          type: string
          format: date-time
          nullable: true
      required: [id, status, diff, rationale, tests_suggested, rollback_plan, created_at]
    PolicyProposalListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PolicyProposalItem"
      required: [items]
...